Extract functionality into custom hooks
