+++ packages/studio/src/application/modules/animation/ui/AnimationCanvas.tsx
@@ -9,6 +9,8 @@ import {
   useRef,
 } from "react";
 
+import { useAnimationsPlaying } from "@/application/modules/animation/hooks/useAnimationsPlaying";
+import { usePanningAndZoom } from "@/application/modules/animation/hooks/usePanningAndZoom";
 import useEvent from "@/application/state/hooks/useEvent";
 import {
   useControlValues,
@@ -16,14 +18,10 @@ import {
   useMutationValues,
 } from "@/application/state/ImageControlContext";
 import {
-  useScreenSubscription,
   useScreenTranslation,
   useUpdateScreenTranslation,
 } from "@/application/state/ScreenTranslationContext";
-import {
-  getInitialScale,
-  maxZoomFactor,
-} from "@/infrastructure/webgl/lib/canvas";
+import { getInitialScale } from "@/infrastructure/webgl/lib/canvas";
 
 export type AnimationCanvasProps = {
   image: HTMLImageElement | null;
@@ -44,16 +42,7 @@ export const AnimationCanvas: FC<PropsWithChildren<AnimationCanvasProps>> = ({
   const playerRef = useRef<GeppettoPlayer | null>(null);
   const animationControlsRef = useRef<AnimationControls | null>(null);
   const animationFrameRef = useRef<number | null>(null);
-  const isDraggingRef = useRef(false);
-  const lastMousePosRef = useRef<{ x: number; y: number } | null>(null);
-  const mouseDeltaRef = useRef<{ x: number; y: number }>({ x: 0, y: 0 });
-
   const translation = useScreenTranslation();
-  const updateScreenTranslation = useUpdateScreenTranslation();
-  const subscribeScreenTranslation = useScreenSubscription();
-  const controlValuesRef = useControlValues();
-  const mutationValuesRef = useMutationValues();
-  const subscribe = useControlValueSubscription();
 
   // Initialize WebGL player
   useEffect(() => {
@@ -109,6 +98,7 @@ export const AnimationCanvas: FC<PropsWithChildren<AnimationCanvasProps>> = ({
     };
   }, [image, file]);
 
+  const updateScreenTranslation = useUpdateScreenTranslation();
   // Calculate the scale factor of how the image fits in the canvas
   useEffect(() => {
     if (!containerRef.current || !image) return;
@@ -158,6 +148,9 @@ export const AnimationCanvas: FC<PropsWithChildren<AnimationCanvasProps>> = ({
     }
   );
 
+  const controlValuesRef = useControlValues();
+  const mutationValuesRef = useMutationValues();
+  const subscribe = useControlValueSubscription();
   // Subscribe to control value changes
   useEffect(() => {
     if (!animationControlsRef.current) return;
@@ -210,172 +203,13 @@ export const AnimationCanvas: FC<PropsWithChildren<AnimationCanvasProps>> = ({
     };
   }, [image, updateScreenTranslation]);
 
-  // Subscribe to translation changes and update player viewport
-  useEffect(() => {
-    const unsubscribe = subscribeScreenTranslation((trans) => {
-      if (!animationControlsRef.current) return;
-      animationControlsRef.current.setZoom(trans.zoom);
-      animationControlsRef.current.setPanning(trans.panX, trans.panY);
-      // If the player needs to react to scale, add logic here
-    });
-    return unsubscribe;
-  }, [subscribeScreenTranslation]);
+  const { handleMouseDown, handleMouseMove, handleMouseUp } = usePanningAndZoom(
+    animationControlsRef,
+    containerRef,
+    image
+  );
 
-  // Mouse event handlers for panning
-  const handleMouseDown = useEvent((e: React.MouseEvent<HTMLCanvasElement>) => {
-    if (e.button !== 0) return; // Only left mouse button
-    if (!containerRef.current || !animationControlsRef.current) return;
-
-    const rect = containerRef.current.getBoundingClientRect();
-    const mouseX = e.clientX - rect.left;
-    const mouseY = e.clientY - rect.top;
-
-    isDraggingRef.current = true;
-    lastMousePosRef.current = { x: mouseX, y: mouseY };
-    mouseDeltaRef.current = { x: 0, y: 0 };
-  });
-
-  const handleMouseMove = useEvent((e: React.MouseEvent<HTMLCanvasElement>) => {
-    if (
-      !isDraggingRef.current ||
-      !lastMousePosRef.current ||
-      !animationControlsRef.current ||
-      !containerRef.current
-    )
-      return;
-
-    const rect = containerRef.current.getBoundingClientRect();
-    const mouseX = e.clientX - rect.left;
-    const mouseY = e.clientY - rect.top;
-
-    const totalDeltaX = mouseX - lastMousePosRef.current.x;
-    const totalDeltaY = mouseY - lastMousePosRef.current.y;
-
-    // Calculate incremental delta since last move
-    const deltaX = totalDeltaX - mouseDeltaRef.current.x;
-    const deltaY = totalDeltaY - mouseDeltaRef.current.y;
-    mouseDeltaRef.current = { x: totalDeltaX, y: totalDeltaY };
-
-    // Get current viewport from player
-    const viewport = animationControlsRef.current.getViewport();
-
-    // Player's pan coordinate space: ±1 = half canvas width (independent of zoom/scale)
-    // deltaPan = deltaPixels / (canvasWidth / 2) = deltaPixels * 2 / canvasWidth
-    const newPanX = Math.min(
-      1.0,
-      Math.max(-1.0, viewport.panX + (deltaX * 2) / rect.width)
-    );
-
-    const newPanY = Math.min(
-      1.0,
-      Math.max(-1.0, viewport.panY - (deltaY * 2) / rect.height)
-    );
-
-    // Update player directly
-    animationControlsRef.current.setPanning(newPanX, newPanY);
-
-    // Sync with context
-    updateScreenTranslation((trans) => ({
-      ...trans,
-      panX: newPanX,
-      panY: newPanY,
-    }));
-  });
-
-  const handleMouseUp = useEvent(() => {
-    isDraggingRef.current = false;
-    lastMousePosRef.current = null;
-    mouseDeltaRef.current = { x: 0, y: 0 };
-  });
-
-  const handleWheel = useEvent((e: React.WheelEvent<HTMLCanvasElement>) => {
-    e.preventDefault();
-    if (!animationControlsRef.current || !containerRef.current) return;
-
-    const rect = containerRef.current.getBoundingClientRect();
-    const mouseX = e.clientX - rect.left;
-    const mouseY = e.clientY - rect.top;
-
-    // Get current viewport
-    const viewport = animationControlsRef.current.getViewport();
-    const maxZoom = maxZoomFactor(image);
-
-    // Calculate new zoom
-    const newZoom = Math.min(
-      maxZoom,
-      Math.max(0.1, viewport.zoom - e.deltaY / 100)
-    );
-
-    if (newZoom === viewport.zoom) return;
-
-    // Convert mouse position to normalized coordinates relative to canvas center
-    // Canvas center is (0, 0), edges are at (-1, -1) to (1, 1)
-    const normalizedMouseX = (mouseX / rect.width) * 2 - 1;
-    const normalizedMouseY = 1 - (mouseY / rect.height) * 2;
-
-    // Calculate the position in the zoomed space before zoom change
-    // The mouse points to a specific location in the image coordinate space
-    const imagePointX = (normalizedMouseX - viewport.panX) / viewport.zoom;
-    const imagePointY = (normalizedMouseY - viewport.panY) / viewport.zoom;
-
-    // After zoom, we want the same image point to be under the mouse
-    // So we solve for the new pan: normalizedMouse = newPan + imagePoint * newZoom
-    const newPanX = normalizedMouseX - imagePointX * newZoom;
-    const newPanY = normalizedMouseY - imagePointY * newZoom;
-
-    // Update player directly
-    animationControlsRef.current.setZoom(newZoom);
-    animationControlsRef.current.setPanning(newPanX, newPanY);
-
-    // Sync with context
-    updateScreenTranslation((trans) => ({
-      ...trans,
-      zoom: newZoom,
-      panX: newPanX,
-      panY: newPanY,
-    }));
-  });
-
-  // Attach wheel event with passive: false to allow preventDefault
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-
-    const wheelHandler = (e: WheelEvent) => {
-      handleWheel(e as unknown as React.WheelEvent<HTMLCanvasElement>);
-    };
-
-    container.addEventListener("wheel", wheelHandler, { passive: false });
-    return () => container.removeEventListener("wheel", wheelHandler);
-  }, [handleWheel]);
-
-  const animationsPlayingRef = useRef<string[]>([]);
-  useEffect(() => {
-    if (!animationControlsRef.current) return;
-    if (
-      JSON.stringify(animationsPlayingRef.current) ===
-      JSON.stringify(animationsPlaying)
-    )
-      return;
-
-    // Animations to start
-    const toStart = animationsPlaying
-      .filter((id) => !animationsPlayingRef.current.includes(id))
-      .map((id) => file.animations[id].name);
-
-    // Animations to stop
-    const toStop = animationsPlayingRef.current
-      .filter((id) => !animationsPlaying.includes(id))
-      .map((id) => file.animations[id].name);
-
-    toStart.forEach((animationName) => {
-      animationControlsRef.current?.startAnimation(animationName);
-    });
-    toStop.forEach((animationName) => {
-      animationControlsRef.current?.stopAnimation(animationName);
-    });
-    animationsPlayingRef.current = animationsPlaying;
-  }, [animationsPlaying, file.animations]);
+  useAnimationsPlaying(animationControlsRef, file, animationsPlaying);
 
   return (
     <div
